Types Used to Define a Model

Hephaestus.Model — Type

struct Model

A type representing the finite element model of a structure of interest.

nodes: Ordered dictionary that stores the nodes of the model

materials: Ordered dictionary that stores the materials of the model

sections: Ordered dictionary that stores the materials of the model

elements: Ordered dictionary that stores the elements of the model

supports: Ordered dictionary that stores the supports of the model

conc_loads: Ordered dictionary that stores the concentrated loads of the model

dist_loads: Ordered dictionary that stores the distributed loads of the model

p_l: Ordered dictionary that stores the fixed-end forces in the local coordinate system

p_g: Ordered dictionary that stores the fixed-end forces in the global coordinate system

Hephaestus.Node — Type

struct Node

A type representing a node in a finite element model.

x: $x$-coordinate

y: $y$-coordinate

z: $z$-coordinate

Hephaestus.Material — Type

struct Material

A type representing a material in a finite element model.

E: Young's modulus

ν: Poisson's ratio

ρ: Density

Hephaestus.Section — Type

struct Section

A type representing a section in a finite element model.

A: Cross-sectional area

I_zz: Moment of inertia about the local $z$-axis

I_yy: Moment of inertia about the local $y$-axis

J: Polar moment of inertia

Hephaestus.Element — Type

struct Element

A type representing an element in the finite element model.

node_i_tag: Unique identifier of the node ($i$)

x_i: $x$-coordinate of the node ($i$)

y_i: $y$-coordinate of the node ($i$)

z_i: $z$-coordinate of the node ($i$)

node_j_tag: Unique identifier of the node ($j$)

x_j: $x$-coordinate of the node ($j$)

y_j: $y$-coordinate of the node ($j$)

z_j: $z$-coordinate of the node ($j$)

material_tag: Unique identifier of the material

E: Young's modulus

ν: Poisson's ratio

ρ: Density

section_tag: Unique identifier of the section

A: Cross-sectional area

I_zz: Moment of inertia about the local $z$-axis

I_yy: Moment of inertia about the local $x$-axis

J: Polar moment of inertia

ω: Angle that defines the orientation of the local coordinate system of the element

releases_i: DOF releases at the node ($i$)

releases_j: DOF releases at the node ($j$)

L: Length of the element

γ: Local-to-global sub-transformation matrix

Γ: Local-to-global transformation matrix

k_e_l: Element elastic stiffness matrix in the local coordinate system

k_e_g: Element elastic stiffness matrix in the global coordinate system

k_g_l: Element geometric stiffness matrix in the local coordinate system

k_g_g: Element geometric stiffness matrix in the global coordinate system

m_l: Element mass matrix in the local coordinate system

m_g: Element mass matrix in the global coordinate system

Types Used to Perform Analyses of Different Types and Store the Results

Hephaestus.LinearElasticAnalysis — Type

struct LinearElasticAnalysis

A type representing the (geometrically) linear (materially) elastic analysis.

Hephaestus.LinearElasticAnalysisCache — Type

struct LinearElasticAnalysisCache

A type used to store the results of linear elastic analysis.

Hephaestus.NonlinearElasticAnalysis — Type

struct NonlinearElasticAnalysis

A type representing the (geometrically) nonlinear (materially) elastic analysis.

Hephaestus.NonlinearElasticAnalysisCache — Type

struct NonlinearElasticAnalysisCache

A type used to store the results of nonlinear elastic analysis.

Hephaestus.ElasticBucklingAnalysis — Type

struct ElasticBucklingAnalysis

A type representing the elastic buckling analysis.

Hephaestus.ElasticBucklingAnalysisCache — Type

struct ElasticBucklingAnalysisCache

A type used to store the results of elastic buckling analysis.

Hephaestus.FreeVibrationAnalysis — Type

struct LinearElasticAnalysis

A type representing the free vibration analysis.

Hephaestus.FreeVibrationAnalysisCache — Type

struct LinearElasticAnalysisCache

A type used to store the results of free vibration analysis.

Functions Used to Define a Model

Hephaestus.add_node! — Function

add_node!(model, tag, x, y, z)

Add a node to the model.

Hephaestus.add_material! — Function

add_material!(model, tag, E, ν, ρ)

Add a material to the model.

Hephaestus.add_section! — Function

add_section!(model, tag, A, I_zz, I_yy, J)

Add a section to the model.

Hephaestus.add_element! — Function

add_element!(model, tag, node_i_tag, node_j_tag, section_tag, material_tag)

Add an element to the model.

Hephaestus.add_support! — Function

add_support!(model, tag, u_x, u_y, u_z, θ_x, θ_y, θ_z)

Add a support to the model.

Hephaestus.add_conc_load! — Function

add_conc_load!(model, tag, F_x, F_y, F_z, M_x, M_y, M_z)

Add a concentrated load to the model.

Hephaestus.add_dist_load! — Function

add_dist_load!(model, tag, q_x, q_y, q_z)

Add a distributed load to the model.

Functions Used to Perform Analyses of Different Types and Extract the Results

Hephaestus.solve — Function

solve(model::Model, analysis::AbstractAnalysisType)

Solve the model using the specified analysis type.

Hephaestus.get_node_u_g — Function

get_node_u_g(solution_cache::AbstractAnalysisCache, tag::Int)

Extracts the displacement vector of a node in the global coordinate system.

Hephaestus.get_element_u_l — Function

get_element_u_l(model::Model, solution_cache::AbstractAnalysisCache, tag::Int)

Extracts the element displacement vector in the local coordinate system.

Hephaestus.get_element_f_l — Function

get_element_f_l(model::Model, solution_cache::AbstractAnalysisCache, tag::Int)

Extract the element force vector in the local coordinate system.

Functions Used to Plot a Model and the Results of Analyses of Different Types

Hephaestus.plotmodel — Function

plotmodel(model::Model, [options])

Plots a model of a structure of interest (Model) into a new Makie.jl scene.

Plotting Options

Nodes

Option	Description	Default

node_visible	Whether to show the nodes	true

node_color	Color of the nodes	:red

node_marker	Marker of the nodes	:circle

node_markersize	Size of the nodes	6

node_strokecolor	Stroke color of the nodes	:black

node_strokewidth	Stroke width of the nodes	1

Node Labels

Option	Description	Default

node_label_visible	Whether to label the nodes	true

align_node_labels	Alignment of the node labels	(:center, :bottom)

node_label_color	Color of the node labels	:red

Elements

Option	Description	Default

element_visible	Whether to show the elements	true

element_color	Color of the elements	:black

element_linestyle	Line style of the elements	:solid

element_linewidth	Line width of the elements	1

Element Labels

Option	Description	Default

element_label_visible	Whether to label the elements	true

align_element_labels	Alignment of the element labels	(:center, :bottom)

element_label_color	Color of the element labels	:black

Supports

Option	Description	Default

support_visible	Whether to show the supports	true

support_color	Color of the supports	:green

Hephaestus.plotmodel! — Function

plotmodel!(model::Model, [options])

Plots a model of a structure of interest (Model) into an existing scene. The plotting options are the same as in plotmodel() function.

Hephaestus.plotsolution — Function

plotsolution(model::Model, solution::AbstractSolutionCache, [options])

Plots undisplaced and displaced shapes of the model of a structure of interest (Model) into a new Makie.jl scene.

Plotting Options

Option	Description	Default

scale	Scale factor	1

Nodes of the Undisplaced Shape

Option	Description	Default

undisp_node_visible	Whether to show the nodes	true

undisp_node_color	Color of the nodes	:steelblue

undisp_node_marker	Marker of the nodes	:circle

undisp_node_markersize	Size of the nodes	6

undisp_node_strokecolor	Stroke color of the nodes	:black

undisp_node_strokewidth	Stroke width of the nodes	1

Node Labels of the Undisplaced Shape

Option	Description	Default

undisp_node_label_visible	Whether to label the nodes	true

undisp_node_label_align	Alignment of the node labels	(:center, :bottom)

undisp_node_label_color	Color of the node labels	:steelblue

Elements of the Undisplaced Shape

Option	Description	Default

undisp_element_visible	Whether to show the elements	true

undisp_element_color	Color of the elements	:black

undisp_element_linestyle	Line style of the elements	:dash

undisp_element_linewidth	Line width of the elements	1

Nodes of the Displaced Shape

Option	Description	Default

disp_node_visible	Whether to show the nodes	true

disp_node_color	Color of the nodes	:crimson

disp_node_marker	Marker of the nodes	:circle

disp_node_markersize	Size of the nodes	6

disp_node_strokecolor	Stroke color of the nodes	:black

disp_node_strokewidth	Stroke width of the nodes	1

Node Labels of the Displaced Shape

Option	Description	Default

disp_node_label_visible	Whether to label the nodes	true

disp_node_label_align	Alignment of the node labels	(:center, :top)

disp_node_label_color	Color of the node labels	:crimson

Elements of the Displaced Shape

Option	Description	Default

disp_element_visible	Whether to show the elements	true

disp_element_color	Color of the elements	:black

disp_element_linestyle	Line style of the elements	:solid

disp_element_linewidth	Line width of the elements	1

Hephaestus.plotsolution! — Function

plotsolution!(model::Model, [options])

Plots undisplaced and displaced shapes of the model of a structure of interest (Model) into an existing scene. The plotting options are the same as in plotsolution() function.