Types

Hephaestus.Model — Type

struct Model

A type representing the finite element model of a structure of interest.

Hephaestus.Node — Type

struct Node

A type representing a node in the finite element model of a structure of interest.

Hephaestus.Material — Type

struct Material

A type representing a material in the finite element model of a structure of interest.

Hephaestus.Section — Type

struct Section

A type representing a section in the finite element model of a structure of interest.

Hephaestus.Element — Type

struct Element

A type representing an element in the finite element in a structure of interest.

Functions

Hephaestus.add_node! — Function

add_node!(model, ID, x, y, z)

Adds a new node to the model.

Hephaestus.del_node! — Function

del_node!(model, ID)

Deletes a node from the model.

Hephaestus.add_material! — Function

add_material!(model, ID, E, υ)

Adds a new material to the model.

Hephaestus.del_material! — Function

del_material!(model, ID)

Deletes a material from the model.

Hephaestus.add_section! — Function

add_section!(model, ID, A, I_zz, I_yy, J)

Adds a new section to the model.

Hephaestus.del_section! — Function

del_section!(model, ID)

Deletes a section from the model.

Hephaestus.add_element! — Function

add_element!(model, ID, node_i_ID, node_j_ID, material_ID, section_ID; 
    ω = 0, 
    releases_i = [false, false, false, false, false, false],
    releases_j = [false, false, false, false, false, false])

Adds a new element to the model.

Hephaestus.del_element! — Function

del_element!(model, ID)

Deletes an element from the model.

Hephaestus.add_support! — Function

add_support!(model, ID, u_x, u_y, u_z, θ_x, θ_y, θ_z)

Adds a new support to the model.

Hephaestus.del_support! — Function

del_support!(model, ID)

Deletes a support from the model.

Hephaestus.add_conc_load! — Function

add_conc_load!(model, ID, F_x, F_y, F_z, M_x, M_y, M_z)

Adds a new concentrated load to the model.

Hephaestus.del_conc_load! — Function

del_conc_load!(model, ID)

Deletes a concentrated load from the model.

Hephaestus.add_dist_load! — Function

add_dist_load!(model, ID, q_x, q_y, q_z)

Adds a new distributed load to the model.

Hephaestus.del_dist_load! — Function

del_dist_load!(model, ID)

Deletes a distributed load from the model.