API

Hephaestus.Model — Type

struct Model

A type representing the finite element model of a structure of interest.

nodes: Nodes of the model

sections: Sections of the model

materials: Materials of the model

elements: Elements of the model

concloads: Concentrated loads of the model

distloads: Distribution loads of the model

Hephaestus.Node — Type

struct Node

A type representing a node in a finite element model.

ID: Identification tag

x: $x$-coordinate

y: $y$-coordinate

z: $z$-coordinate

u_x: Is the node restrained against translation along the global $x$-axis?

u_y: Is the node restrained against translation along the global $y$-axis?

u_z: Is the node restrained against translation along the global $z$-axis?

θ_x: Is the node restrained against rotation about the global $x$-axis?

θ_y: Is the node restrained against rotation about the global $y$-axis?

θ_z: Is the node restrained against rotation about the global $z$-axis?

state: Current state

Hephaestus.Material — Type

struct Material

A type representing a material in a finite element model.

ID: Identification tag

E: Young's modulus, $E$

ν: Poisson's ratio, $\nu$

ρ: Density, $\rho$

Hephaestus.Section — Type

struct Section

A type representing a section in a finite element model.

ID: Identification tag

A: Cross-sectional area, $A$

I_zz: Moment of inertia about the local $z$-axis, $I_{zz}$

I_yy: Moment of inertia about the local $y$-axis, $I_{yy}$

J: Torsional constant, $J$

Hephaestus.Element — Type

struct Element

A type representing an element in the finite element model.

ID: Identification tag

node_i: Node ($i$) of the element

node_j: Node ($j$) of the element

section: Section attached to the element

material: Material attached to the element

ω: Orientation angle of the element

releases_i: End moment releases at node ($i$) of the element

releases_j: End moment releases at node ($j$) of the element

state: Current state

Hephaestus.LinearElasticAnalysis — Type

struct LinearElasticAnalysis <: AbstractAnalysisType

A type representing (geometrically) linear (materially) elastic analysis.

Hephaestus.NonlinearElasticAnalysis — Type

struct NonlinearElasticAnalysis <: AbstractAnalysisType

A type representing (geometrically) nonlinear (materially) elastic analysis.

Hephaestus.ElasticBucklingAnalysis — Type

struct ElasticBucklingAnalysis <: AbstractAnalysisType

A type representing the elastic buckling analysis.

To perform an elastic buckling analysis, use the following command:

solution = solve(model, ElasticBucklingAnalysis())

Hephaestus.ElasticBucklingAnalysisCache — Type

struct ElasticBucklingAnalysisCache <: AbstractSolutionCache

A type used to store the results of elastic buckling analysis.

Hephaestus.FreeVibrationAnalysis — Type

struct FreeVibrationAnalysis <: AbstractAnalysisType

A type representing the free vibration analysis.

To perform an free vibration analysis, use the following command:

solution = solve(model, FreeVibrationAnalysis())

Hephaestus.FreeVibrationAnalysisCache — Type

struct FreeVibrationAnalysisCache <: AbstractSolutionCache

A type used to store the results of free vibration analysis.

Hephaestus.node! — Function

node!(model::Model, ID::Int,
    x::Real, y::Real, z::Real;
    u_x::Bool = false, u_y::Bool = false, u_z::Bool = false,
    θ_x::Bool = false, θ_y::Bool = false, θ_z::Bool = false)::Model

Add a node to a finite element model.

Hephaestus.section! — Function

section!(model::Model, ID::Int,
    A::Real, I_zz::Real, I_yy::Real, J::Real)::Model

Add a section to a finite element model.

Hephaestus.material! — Function

material!(model::Model, ID::Int,
    E::Real, ν::Real, ρ::Real)::Model

Add a material to a finite element model.

Hephaestus.element! — Function

element!(model::Model, ID::Int,
    node_i_ID  ::Int,
    node_j_ID  ::Int,
    section_ID ::Int,
    material_ID::Int;
    ω          ::Real = 0,
    releases_i ::Vector{<:Bool} = [false, false, false],
    releases_j ::Vector{<:Bool} = [false, false, false])::Model

Add an element to a finite element model.

Hephaestus.concload! — Function

concload!(model::Model, ID::Int,
    F_x::Real, F_y::Real, F_z::Real,
    M_x::Real, M_y::Real, M_z::Real)::Model

Applies a concentrated load to a node with a specified ID.

Hephaestus.distload! — Function

distload!(model::Model, ID::Int,
    w_x::Real, w_y::Real, w_z::Real)::Model

Applies a distributed load to an element with a specified ID.

Hephaestus.solve! — Function

solve!(model::Model, analysistype::AbstractAnalysisType)::Model

Function used to perform (geometrically) linear and nonlinear and (materially) elastic and inelastic analyses.

Note

This function is mutating and will update the state of the model.

Hephaestus.solve — Function

solve(model::Model, analysistype::AbstractAnalysisType)::AbstractSolutionCache

Function used to perform elastic buckling and free vibration analyses.

Note

This function is nonmutating and will not update the state of the model. Instead, it return a solution cache object that can be used to extract the results.

Postprocessing functions

Hephaestus.getnodaldisplacements — Function

getnodaldisplacements(model::Model, ID::Int)

Extracts the displacement vector of a node of interest.

Hephaestus.getnodalreactions — Function

getnodalreactions(model::Model, ID::Int)

Extracts the reaction vector of a node of interest from the solution cache.

Hephaestus.getelementdisplacements — Function

getelementdisplacements(model::Model, ID::Int)

Extracts the displacement vector of an element of interest in its local coordinate system.

Hephaestus.getelementforces — Function

getelementforces(model::Model, ID::Int)

Extracts the internal force vector of an element of interest in its local coordinate system.

Plotting functions

Hephaestus.plotmodel — Function

plotmodel(model::Model, [options])

Plots a model of a structure of interest (Model) into a new Makie.jl scene.

Hephaestus.plotmodel! — Function

plotmodel!(scene::Scene, model::Model, [options])

Plots a model of a structure of interest (Model) into an existing Makie.jl scene.