API

Types

Hephaestus.Model — Type

struct Model

A type that represents the FE model of a structure.

To initialize a new model, use the Model() constructor.

Hephaestus.Node — Type

struct Node

A type that represents a node in the FE model of a structure.

This type should not be used directly. Instead, use the add_node! function to add a node to the model. To remove a node from the model, use the del_node! function.

Hephaestus.Material — Type

struct Material

A type that represents a material in the FE model of a structure.

This type should not be used directly. Instead, use the add_material! function to add a material to the model. To remove a material from the model, use the del_material! function.

Hephaestus.Section — Type

struct Section

A type that represents a section in the FE model of a structure.

This type should not be used directly. Instead, use the add_section! function to add a section to the model. To remove a section from the model, use the del_section! function.

Hephaestus.Element — Type

struct Element

A type that represents an element in the FE model of a structure.

This type should not be used directly. Instead, use the add_element! function to add an element to the model. To remove an element from the model, use the del_element! function.

Functions

Hephaestus.add_node! — Function

Model()

add_node!(model, ID, 
    x, y, z)

Adds a node to the model.

Hephaestus.del_node! — Function

del_node!(model, ID)

Deletes a node from the model.

Hephaestus.add_material! — Function

add_material!(model, ID, 
    E, ν)

Adds a material to the model.

Hephaestus.del_material! — Function

del_material!(model, ID)

Deletes a material from the model.

Hephaestus.add_section! — Function

add_section!(model, ID, 
    A, I_zz, I_yy, J)

Adds a section to the model.

Hephaestus.del_section! — Function

del_section!(model, ID)

Deletes a section from the model.

Hephaestus.add_element! — Function

add_element!(model, ID, 
    node_i_ID, node_j_ID, material_ID, section_ID;
    ω = 0, releases = falses(12))

Adds an element to the model.

Hephaestus.del_element! — Function

del_element!(model, ID)

Deletes an element from the model.

Hephaestus.add_nodal_load! — Function

add_nodal_load!(model, ID, 
    F_x, F_y, F_z, 
    M_x, M_y, M_z)

Applies nodal loads to a node in the model.

Hephaestus.del_nodal_load! — Function

del_nodal_load!(model, ID)

Removes nodal loads from a node in the model.

Hephaestus.reset_model! — Function

reset_model!(model)

Completely resets the model to a clean state.