Types Used to Define a Model

Hephaestus.Model — Type

struct Model

A type representing the finite element model of a structure of interest.

nodes: Ordered dictionary that stores the nodes of the model

materials: Ordered dictionary that stores the materials of the model

sections: Ordered dictionary that stores the sections of the model

elements: Ordered dictionary that stores the elements of the model

springs: Ordered dictionary that stores the springs of the model

supports: Ordered dictionary that stores the supports of the model

conc_loads: Ordered dictionary that stores the concentrated loads of the model

dist_loads: Ordered dictionary that stores the distributed loads of the model

p_l: Ordered dictionary that stores the element fixed-end forces in the local coordinate system

p_g: Ordered dictionary that stores the element fixed-end forces in the global coordinate system

Hephaestus.Node — Type

struct Node

A type representing a node in the model of a structure of interest.

ID: Unique identifier

x: $x$-coordinate

y: $y$-coordinate

z: $z$-coordinate

Hephaestus.Material — Type

struct Material

A type representing a material in the model of a structure of interest.

ID: Unique identifier

E: Young's modulus, $E$

ν: Poisson's ratio, $u$

ρ: Density, $\rho$

Hephaestus.Section — Type

struct Section

A type representing a section in the model of a structure of interest.

ID: Unique identifier

A: Cross-sectional area, $A$

I_zz: Moment of inertia about the local $z$-axis, $I_{zz}$

I_yy: Moment of inertia about the local $y$-axis, $I_{yy}$

J: Polar moment of inertia about the local $x$-axis, $J$

Hephaestus.Element — Type

struct Element

A type representing an element in the model of a structure of interest.

ID: Unique identifier

node_i_ID: Unique identifier of the node $i$ of the element

node_j_ID: Unique identifier of the node $j$ of the element

material_ID: Unique identifier of the material of the element

section_ID: Unique identifier of the section of the element

releases_i: DOF releases at the node $i$

releases_j: DOF releases at the node $j$

ω: Angle that defines the orientation of the element's local coordinate system, $\omega$

x_i: $x$-coordinate of the node $i$, $x_{i}$

y_i: $y$-coordinate of the node $i$, $y_{i}$

z_i: $z$-coordinate of the node $i$, $z_{i}$

x_j: $x$-coordinate of the node $j$, $x_{j}$

y_j: $y$-coordinate of the node $j$, $y_{j}$

z_j: $z$-coordinate of the node $j$, $z_{j}$

E: Young's modulus, $E$

ν: Poisson's ratio, $\nu$

ρ: Density, $\rho$

A: Cross-sectional area, $A$

I_zz: Moment of inertia about the local $z$-axis, $I_{zz}$

I_yy: Moment of inertia about the local $y$-axis, $I_{yy}$

J: Polar moment of inertia about the local $x$-axis, $J$

L: Length of the element, $L$

γ: Local-to-global sub-transformation matrix, $\gamma$

Γ: Local-to-global transformation matrix, $\Gamma$

k_e_l: Elastic stiffness matrix in the local coordinate system of the element, $k_{e, l}$

k_e_g: Elastic stiffness matrix in the global coordinate system, $k_{e, g}$

k_g_l: Geometric stiffness matrix in the local coordinate system of the element, $k_{g, l}$

k_g_g: Geometric stiffness matrix in the global coordinate system, $k_{g, g}$

m_l: Mass matrix in the local coordinate system of the element, $m_{l}$

m_g: Mass matrix in the global coordinate system, $m_{g}$

Types Used to Perform Analyses of Different Types and Store the Results

Hephaestus.O1EAnalysis — Type

struct O1EAnalysis

A type that represents the 1st-order (O1) elastic (E) analysis.

Hephaestus.O1ESolutionCache — Type

struct O1ESolutionCache

A type that stores the results of the 1st-order elastic analysis.

Hephaestus.O2EAnalysis — Type

struct O2EAnalysis

A type that represents the 2nd-order (O2) elastic (E) analysis.

Hephaestus.O2ESolutionCache — Type

struct O2ESolutionCache

A type that stores the results of the 2nd-order elastic analysis.

Hephaestus.EBAnalysis — Type

struct EBAnalysis

A type that represents the elastic buckling (EB) analysis.

Hephaestus.EBSolutionCache — Type

struct EBSolutionCache

A type that stores the results of the elastic buckling analysis.

Hephaestus.FVAnalysis — Type

struct FVAnalysis

A type that represents the free vibration (FV) analysis.

Hephaestus.FVSolutionCache — Type

struct FVSolutionCache

A type that stores the results of the free vibration analysis.

Functions Used to Define a Model

Hephaestus.add_node! — Function

add_node!(model::Model, ID::Int, 
    x::Real, y::Real, z::Real)

Adds a node to the model.

Hephaestus.del_node! — Function

del_node!(model::Model, ID::Int)

Deletes a node from the model and all related elements, supports, and concentrated loads.

Hephaestus.add_material! — Function

add_material!(model::Model, ID::Int, 
    E::Real, ν::Real, ρ::Real)

Adds a material to the model.

Hephaestus.del_material! — Function

del_material!(model::Model, ID::Int)

Deletes a material from the model and all related elements.

Hephaestus.add_section! — Function

add_section!(model::Model, ID::Int, 
    A::Real, I_zz::Real, I_yy::Real, J::Real)

Adds a section to the model.

Hephaestus.del_section! — Function

del_section!(model::Model, ID::Int)

Deletes a section from the model and all related elements.

Hephaestus.add_element! — Function

add_element!(model::Model, ID::Int,
    node_i_ID::Int, node_j_ID::Int, material_ID::Int, section_ID::Int;
    releases_i::Vector{Bool} = [false, false, false, false, false, false],
    releases_j::Vector{Bool} = [false, false, false, false, false, false],
    ω::Real = 0)

Adds an element to the model.

Hephaestus.del_element! — Function

del_element!(model::Model, ID::Int)

Deletes an element from the model and all related distributed loads.

Hephaestus.add_support! — Function

add_support!(model::Model, ID::Int, 
    u_x::Bool, u_y::Bool, u_z::Bool, 
    θ_x::Bool, θ_y::Bool, θ_z::Bool)

Adds a support to the model.

Hephaestus.del_support! — Function

del_support!(model::Model, ID::Int)

Deletes a support from the model.

Hephaestus.add_conc_load! — Function

add_conc_load!(model::Model, ID::Int,
    F_x::Real, F_y::Real, F_z::Real,
    M_x::Real, M_y::Real, M_z::Real)

Adds a concentrated load to the model.

Hephaestus.del_conc_load! — Function

del_conc_load!(model::Model, ID::Int)

Deletes a concentrated load from the model.

Hephaestus.add_dist_load! — Function

add_dist_load!(model::Model, ID::Int,
    q_x::Real, q_y::Real, q_z::Real)

Adds a distributed load to the model.

Hephaestus.del_dist_load! — Function

del_dist_load!(model::Model, ID::Int)

Deletes a distributed load from the model.

Functions Used to Perform Analyses of Different Types and Extract the Results

Hephaestus.solve — Function

solve(model::Model, analysis::AbstractAnalysisType)

The main function that performs the analysis of your choice on the model.

Hephaestus.get_node_u_g — Function

get_node_u_g(solution_cache::AbstractSolutionCache, ID::Int)

Extracts the displacement vector of a node in the global coordinate system.

Hephaestus.get_element_u_l — Function

get_element_u_l(model::Model, solution_cache::AbstractSolutionCache, ID::Int)

Extracts the element displacement vector in the local coordinate system.

Hephaestus.get_element_f_l — Function

get_element_f_l(model::Model, solution_cache::AbstractSolutionCache, ID::Int)

Extract the element force vector in the local coordinate system.

Functions Used to Plot a Model and the Results of Analyses of Different Types

Hephaestus.plotmodel — Function

plotmodel(model::Model, [options])

Plots a model of a structure of interest (Model) into a new Makie.jl scene.

Plotting Options

Nodes

Option	Description	Default

node_visible	Whether to show the nodes	true

node_color	Color of the nodes	:red

node_marker	Marker of the nodes	:circle

node_markersize	Size of the nodes	6

node_strokecolor	Stroke color of the nodes	:black

node_strokewidth	Stroke width of the nodes	1

Node Labels

Option	Description	Default

node_label_visible	Whether to label the nodes	true

align_node_labels	Alignment of the node labels	(:center, :bottom)

node_label_color	Color of the node labels	:red

Elements

Option	Description	Default

element_visible	Whether to show the elements	true

element_color	Color of the elements	:black

element_linestyle	Line style of the elements	:solid

element_linewidth	Line width of the elements	1

Element Labels

Option	Description	Default

element_label_visible	Whether to label the elements	true

align_element_labels	Alignment of the element labels	(:center, :bottom)

element_label_color	Color of the element labels	:black

Supports

Option	Description	Default

support_visible	Whether to show the supports	true

support_color	Color of the supports	:green

Hephaestus.plotmodel! — Function

plotmodel!(model::Model, [options])

Plots a model of a structure of interest (Model) into an existing scene. The plotting options are the same as in plotmodel() function.

Hephaestus.plotsolution — Function

plotsolution(model::Model, solution::AbstractSolutionCache, [options])

Plots undisplaced and displaced shapes of the model of a structure of interest (Model) into a new Makie.jl scene.

Plotting Options

Option	Description	Default

scale	Scale factor	1

Nodes of the Undisplaced Shape

Option	Description	Default

undisp_node_visible	Whether to show the nodes	true

undisp_node_color	Color of the nodes	:steelblue

undisp_node_marker	Marker of the nodes	:circle

undisp_node_markersize	Size of the nodes	6

undisp_node_strokecolor	Stroke color of the nodes	:black

undisp_node_strokewidth	Stroke width of the nodes	1

Node Labels of the Undisplaced Shape

Option	Description	Default

undisp_node_label_visible	Whether to label the nodes	true

undisp_node_label_align	Alignment of the node labels	(:center, :bottom)

undisp_node_label_color	Color of the node labels	:steelblue

Elements of the Undisplaced Shape

Option	Description	Default

undisp_element_visible	Whether to show the elements	true

undisp_element_color	Color of the elements	:black

undisp_element_linestyle	Line style of the elements	:dash

undisp_element_linewidth	Line width of the elements	1

Nodes of the Displaced Shape

Option	Description	Default

disp_node_visible	Whether to show the nodes	true

disp_node_color	Color of the nodes	:crimson

disp_node_marker	Marker of the nodes	:circle

disp_node_markersize	Size of the nodes	6

disp_node_strokecolor	Stroke color of the nodes	:black

disp_node_strokewidth	Stroke width of the nodes	1

Node Labels of the Displaced Shape

Option	Description	Default

disp_node_label_visible	Whether to label the nodes	true

disp_node_label_align	Alignment of the node labels	(:center, :top)

disp_node_label_color	Color of the node labels	:crimson

Elements of the Displaced Shape

Option	Description	Default

disp_element_visible	Whether to show the elements	true

disp_element_color	Color of the elements	:black

disp_element_linestyle	Line style of the elements	:solid

disp_element_linewidth	Line width of the elements	1

Hephaestus.plotsolution! — Function

plotsolution!(model::Model, [options])

Plots undisplaced and displaced shapes of the model of a structure of interest (Model) into an existing scene. The plotting options are the same as in plotsolution() function.