API

Types Used to Define a Model

Hephaestus.Model — Type

struct Model

A type representing the finite element model of a structure of interest.

name: Name of the model

nodes: Nodes of the model

sections: Sections of the model

materials: Materials of the model

elements: Elements of the model

concloads: Concentrated loads of the model

distloads: Distribution loads of the model

Hephaestus.Node — Type

struct Node

A type representing a node in a finite element model.

ID: Identification tag

x: $x$-coordinate

y: $y$-coordinate

z: $z$-coordinate

u_x: Is the node restrained agaist translation in the global $x$-direction?

u_y: Is the node restrained agaist translation in the global $y$-direction?

u_z: Is the node restrained agaist translation in the global $z$-direction?

θ_x: Is the node restrained agaist rotation about the global $x$-axis?

θ_y: Is the node restrained agaist rotation about the global $y$-axis?

θ_z: Is the node restrained agaist rotation about the global $z$-axis?

Hephaestus.Material — Type

struct Material

A type representing a material in a finite element model.

ID: Identification tag

E: Young's modulus, $E$

ν: Poisson's ratio, $\nu$

ρ: Density, $ρ$

Hephaestus.Section — Type

struct Section

A type representing a section in a finite element model.

ID: Identification tag

A: Cross-sectional area, $A$

I_zz: Moment of inertia about the local $z$-axis, $I_{zz}$

I_yy: Moment of inertia about the local $y$-axis, $I_{yy}$

J: Torsional constant, $J$

Hephaestus.Element — Type

struct Element

A type representing an element in the finite element model.

ID: Identification tag

node_i: Node ($i$)

node_j: Node ($j$)

section: Section

material: Material

ω: Orientation angle, $\omega$

releases_i: End releases at node ($i$)

releases_j: End releases at node ($j$)

L: Element length, $L$

γ: Subtransformation matrix, $\gamma$

Γ: Transformation matrix, $\Gamma$

k_e_l: Elastic stiffness matrix in the local coordinate system, $k_{el}$

k_e_g: Elastic stiffness matrix in the global coordinate system, $k_{eg}$

k_g_l: Geometric stiffness matrix in the local coordinate system, $k_{gl}$

k_g_g: Geometric stiffness matrix in the global coordinate system, $k_{gg}$

m_l: Mass matrix in the local coordinate system, $m_{l}$

m_g: Mass matrix in the global coordinate system, $m_{g}$

Types Used to Perform Analyses of Different Types and Store the Results

Hephaestus.LinearElasticAnalysis — Type

struct LinearElasticAnalysis

A type representing the (geometrically) linear (materially) elastic analysis.

Hephaestus.LinearElasticAnalysisCache — Type

struct LinearElasticAnalysisCache

A type used to store the results of linear elastic analysis.

Hephaestus.ElasticBucklingAnalysis — Type

struct ElasticBucklingAnalysis

A type representing the elastic buckling analysis.

Hephaestus.ElasticBucklingAnalysisCache — Type

struct ElasticBucklingAnalysisCache

A type used to store the results of elastic buckling analysis.

Hephaestus.FreeVibrationAnalysis — Type

struct LinearElasticAnalysis

A type representing the free vibration analysis.

Hephaestus.FreeVibrationAnalysisCache — Type

struct LinearElasticAnalysisCache

A type used to store the results of free vibration analysis.

Functions Used to Define a Model

Hephaestus.node! — Function

node!(model, ID, 
    x, y, z;
    u_x = false, u_y = false, u_z = false,
    θ_x = false, θ_y = false, θ_z = false)

Add a node to a finite element model.

Hephaestus.section! — Function

section!(model, ID, 
    A, I_zz, I_yy, J)

Add a section to a finite element model.

Hephaestus.material! — Function

material!(model, ID, 
    E, ν, ρ)

Add a material to a finite element model.

Hephaestus.element! — Function

element!(model, ID, 
    node_i_ID, node_j_ID, 
    section_ID, 
    material_ID; 
    ω = 0,
    releases_i = [false, false, false, false, false, false],
    releases_j = [false, false, false, false, false, false])

Add an element to a finite element model.

Hephaestus.concload! — Function

concload!(model, ID, 
    F_x, F_y, F_z, 
    M_x, M_y, M_z)

Applies a concentrated load to a node with a specified ID.

Hephaestus.distload! — Function

distload!(model, ID, 
    w_x, w_y, w_z; 
    cs = :local)

Applies a distributed load to an element with a specified ID.

Functions Used to Perform Analyses of Different Types and Extract the Results

Hephaestus.solve — Function

solve(model::Model, analysistype::AbstractAnalysisType)

Solve the model using the specified analysis type.

Hephaestus.getnodedisp — Function

getnodedisp(model::Model, solution::AbstractSolutionCache, ID::Int)

Extracts the displacement vector of a node of interest from the solution cache.

Hephaestus.getelementdisp_l — Function

getelementdisp_l(model::Model, solution::AbstractSolutionCache, ID::Int)

Extracts the displacement vector of an element of interest in the local coordinate system from the solution cache.

Hephaestus.getelementdisp_g — Function

getelementdisp_g(model::Model, solution::AbstractSolutionCache, ID::Int)

Extracts the displacement vector of an element of interest in the global coordinate system from the solution cache.

Hephaestus.getelementforces_l — Function

getelementforces_l(model::Model, solution::AbstractSolutionCache, ID::Int)

Extracts the internal forces of an element of interest in the local coordinate system from the solution cache.

Hephaestus.getelementforces_g — Function

getelementforces_g(model::Model, solution::AbstractSolutionCache, ID::Int)

Extracts the internal forces of an element of interest in the global coordinate system from the solution cache.

Functions Used to Plot a Model and the Results of Analyses of Different Types

Hephaestus.plotmodel — Function

plotmodel(model::Model, [options])

Plots a model of a structure of interest (Model) into a new Makie.jl scene.

Hephaestus.plotmodel! — Function

plotmodel!(scene::Scene, model::Model, [options])

Plots a model of a structure of interest (Model) into an existing Makie.jl scene.