Hephaestus.Model — Type

Model

A type that represents the finite element model of a structure.

To create a model, use the Model() constructor.

Fields

nodes: Dictionary of nodes in the model.

materials: Dictionary of materials in the model.

sections: Dictionary of sections in the model.

elements: Dictionary of elements in the model.

supports: Dictionary of supports in the model.

concentrated_loads: Dictionary of concentrated loads (acting on nodes) in the model.

distributed_loads: Dictionary of distributed loads (acting on elements) in the model.

p_l: Dictionary of element fixed-end forces in the element's local coordinate system.

p_g: Dictionary of element fixed-end forces in the global coordinate system.

Hephaestus.Node — Type

Node

A type that represents a node in the finite element model of a structure.

Fields

ID: Unique identifier of the node provided by the user

x: $x$-coordinate of the node

y: $y$-coordinate of the node

z: $z$-coordinate of the node

Hephaestus.Material — Type

Material

A type that represents a material in the finite element model of a structure.

Fields

ID: Unique identifier of the material provided by the user

E: Young's modulus of the material, $E$

ν: Poisson's ratio of the material, $\nu$

ρ: Density of the material, $\rho$

Hephaestus.Section — Type

Section

A type that represents a section in the finite element model of a structure.

Fields

ID: Unique identifier of the section provided by the user

A: Cross-sectional area of the section, $A$

I_zz: Moment of inertia about the local $z$-axis of the section, $I_{zz}$

I_yy: Moment of inertia about the local $y$-axis of the section, $I_{yy}$

J: Polar moment of inertia of the section, $J$

Hephaestus.Element — Type

Element

A type that represents an element in the finite element model of a structure.

This type should never be created directly.

Fields

ID: Unique identifier of the element provided by the user

node_i_ID: Unique identifier of the node $i$ of the element provided by the user

node_j_ID: Unique identifier of the node $j$ of the element provided by the user

material_ID: Unique identifier of the material of the element provided by the user

section_ID: Unique identifier of the section of the element provided by the user

x_i: $x$-coordinate of the node $i$ of the element, $x_{i}$

y_i: $y$-coordinate of the node $i$ of the element, $y_{i}$

z_i: $z$-coordinate of the node $i$ of the element, $z_{i}$

x_j: $x$-coordinate of the node $j$ of the element, $x_{j}$

y_j: $y$-coordinate of the node $j$ of the element, $y_{j}$

z_j: $z$-coordinate of the node $j$ of the element, $z_{j}$

E: Young's modulus of the material of the element, $E$

ν: Poisson's ratio of the material of the element, $\nu$

ρ: Density of the material of the element, $\rho$

A: Cross-sectional area of the section of the element, $A$

I_zz: Moment of inertia about the local $z$-axis of the section of the element, $I_{zz}$

I_yy: Moment of inertia about the local $y$-axis of the section of the element, $I_{yy}$

J: Polar moment of inertia of the section of the element, $J$

ω: Angle that defines the orientation of the local coordinate system of the element, $\omega$

releases: Releases at the ends of the element

L: Length of the element, $L$

γ: Global-to-local sub-transformation matrix of the element, $[\gamma]$

T: Global-to-local transformation matrix of the element, $[T]$

k_e_l: Element's elastic stiffness matrix in its local coordinate system, $[k_{e, l}]$

k_g_l: Element's geometric stiffness matrix in its local coordinate system, $[k_{g, l}]$

k_e_g: Element's elastic stiffness matrix in its global coordinate system, $[k_{e, g}]$

k_g_g: Element's geometric stiffness matrix in its global coordinate system, $[k_{g, g}]$